+++ src/main/java/uk/gov/companieshouse/overseasentitiesapi/model/updatesubmission/changelist/additions/IndividualBeneficialOwnerAddition.java
@@ -1,5 +1,8 @@
 package uk.gov.companieshouse.overseasentitiesapi.model.updatesubmission.changelist.additions;
 
+import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import uk.gov.companieshouse.overseasentitiesapi.model.updatesubmission.changelist.commonmodels.PersonName;
 import uk.gov.companieshouse.overseasentitiesapi.model.updatesubmission.changelist.commonmodels.Address;
@@ -13,16 +16,21 @@ public class IndividualBeneficialOwnerAddition extends BeneficialOwnerAddition {
     @JsonProperty("personName")
     private PersonName personName;
 
+    @JsonInclude(NON_NULL)
+    @JsonProperty("residentialAddress")
+    private Address residentialAddress;
+
     @JsonProperty("nationalityOther")
     private String nationalityOther;
 
     @JsonProperty("birthDate")
     private LocalDate birthDate;
 
-    public IndividualBeneficialOwnerAddition(LocalDate actionDate, LocalDate ceasedDate, Address residentialResidence,
+    public IndividualBeneficialOwnerAddition(LocalDate actionDate, LocalDate ceasedDate, Address residentialAddress,
                                              Address serviceAddress, List<String> natureOfControls, boolean isOnSanctionsList) {
-        super(actionDate, ceasedDate, residentialResidence, serviceAddress, natureOfControls, isOnSanctionsList);
+        super(actionDate, ceasedDate, serviceAddress, natureOfControls, isOnSanctionsList);
         setAppointmentType(APPOINTMENT_TYPE);
+        this.residentialAddress = residentialAddress;
     }
 
     public PersonName getPersonName() {
@@ -31,6 +39,14 @@ public class IndividualBeneficialOwnerAddition extends BeneficialOwnerAddition {
 
     public void setPersonName(PersonName personName) {
         this.personName = personName;
+    }
+
+    public Address getResidentialAddress() {
+        return residentialAddress;
+    }
+
+    public void setResidentialAddress(Address residentialAddress) {
+        this.residentialAddress = residentialAddress;
     }
 
     public String getNationalityOther() {
